pyproject.toml
[tool.poetry]
name = "zappi-cake"
version = "0.1.0"
description = "A cake store store api"
authors = ["Leonardo Zappi Ramos <leo.zappi@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.6"
django-rest-framework = "^0.1.0"
djangorestframework = "^3.14.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
